Rename py_invader/old_py_invader to old_py_invader
Changed the old py_invader game to be created in a folder instead.  I want to replace the game with object oriented code, and to finish it this time so that it is playable as an .exe file that can be distributed as a proper game.  This is an exercise from Dusty Phillips book Object Oriented Programming, 3rd edition.  I have to refactor an old program into objects etc. Using the code that I learned.  Very excited about this actually because it will take like 300+ lines of jumbled code and clean that up a bit into modules.